Improve performance of start plan page
Get the count stream in the plans page and pass it to the
start page instead of re-getting it on each plan.
The query is almost the same in terms of performance if we
were to filter by the plan exercises, so its faster to just
do a single grouped query at the top level.